requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
cfgv==3.4.0
coverage==7.12.0
distlib==0.4.0
Django==4.2.8
djangorestframework==3.14.0
filelock==3.20.0
identify==2.6.15
iniconfig==2.3.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
virtualenv==20.35.4
psycopg2==2.9.9 ; platform_machine != "aarch64" and platform_machine != "arm64"
psycopg2-binary==2.9.9 ; platform_machine == "aarch64" or platform_machine == "arm64"
